Extend the BuilderConnection to handle ports
* By the way, the builder is moved to its own file.
