+++ rust/agama-dbus-server/src/network/model.rs
@@ -15,6 +15,7 @@ use std::{
 use thiserror::Error;
 use uuid::Uuid;
 use zbus::zvariant::Value;
+mod builder;
 
 #[derive(Default, Clone)]
 pub struct NetworkState {
@@ -241,52 +242,6 @@ mod tests {
 pub struct Device {
     pub name: String,
     pub type_: DeviceType,
-}
-
-#[derive(Debug, Default)]
-pub struct ConnectionBuilder {
-    id: String,
-    interface: Option<String>,
-    controller: Option<String>,
-    type_: Option<DeviceType>,
-}
-
-impl ConnectionBuilder {
-    pub fn new(id: &str) -> Self {
-        Self {
-            id: id.to_string(),
-            ..Default::default()
-        }
-    }
-
-    pub fn with_interface(mut self, interface: &str) -> Self {
-        self.interface = Some(interface.to_string());
-        self
-    }
-
-    pub fn with_controller(mut self, controller: &str) -> Self {
-        self.controller = Some(controller.to_string());
-        self
-    }
-
-    pub fn with_type(mut self, type_: DeviceType) -> Self {
-        self.type_ = Some(type_);
-        self
-    }
-
-    pub fn build(self) -> Connection {
-        let mut conn = Connection::new(self.id, self.type_.unwrap_or(DeviceType::Ethernet));
-
-        if let Some(interface) = self.interface {
-            conn.set_interface(&interface);
-        }
-
-        if let Some(controller) = self.controller {
-            conn.set_controller(&controller);
-        }
-
-        conn
-    }
 }
 
 /// Represents an available network connection
